implemented cases design by one and implemented the email template to send the token
