+++ src/pages/blog_client/sections/Header.tsx
@@ -1,109 +1,57 @@
-import { FaMicroscope, FaMoon, FaSun } from "react-icons/fa";
-import { Divider, Switch } from "@nextui-org/react";
-import { GiHamburgerMenu } from "react-icons/gi";
-import React, { useContext, useState } from "react";
-import DarkModeContext from "~/context/DarkModeContext";
-import ClientContext from "~/pages/blog_client/context/ClientContext";
-import { ABOUT, CASES } from "~/constants";
-import { useNavigate } from "react-router-dom";
-
+import { BookOpen, Menu, X } from "lucide-react";
+import React, { useState } from "react";
+import { FaMicroscope } from "react-icons/fa";
+import { menuOptions } from "~/constants/options/landing.options";
 function Header() {
-  const [onOpen, setOnOpen] = useState(false);
-  const { setTheme, theme } = useContext(DarkModeContext);
-  const { handleClickOption } = useContext(ClientContext);
-  const handleDarkMode = () => {
-    if (theme === "dark") {
-      setTheme("light");
-    } else {
-      setTheme("dark");
-    }
-  };
-  const navigate = useNavigate();
-  const returnToLanding = () => {
-    navigate("/");
+  const [isMenuOpen, setIsMenuOpen] = useState(false);
+  const [active, setActive] = useState(0);
+
+  const scrollToSection = (sectionId: string) => {
+    document.getElementById(sectionId)?.scrollIntoView({ behavior: "smooth" });
   };
 
   return (
-    <>
+    <header
+      id="hero"
+      className="fixed z-50 flex h-16 w-full items-center bg-purple-950/80 px-4 backdrop-blur-md lg:px-6"
+    >
+      <span
+        className="flex cursor-pointer items-center justify-center"
+        onClick={() => scrollToSection("home")}
+      >
+        <FaMicroscope size={30} className="text-yellow-400" />
+        <h1 className="ml-4 font-bold text-violet-100 md:text-xl lg:text-2xl">Blog de Nandy</h1>
+      </span>
       <nav
-        id="header"
-        className="fixed top-0 z-30 w-full bg-gradient-to-b from-fuchsia-700/80 to-violet-700/80 text-white backdrop-blur dark:from-violet-700/80 dark:to-cyan-700/80"
+        className={`ml-auto flex gap-4 sm:gap-6 ${
+          isMenuOpen ? "flex" : "hidden"
+        } absolute left-0 right-0 top-16 flex-col bg-purple-950 p-4 md:relative md:top-0 md:flex md:flex-row md:bg-transparent md:p-0`}
       >
-        <div className="container mx-auto mt-0 flex w-full flex-wrap items-center justify-between py-3 lg:py-5">
-          <div
-            className="flex cursor-pointer items-center pl-4 active:text-violet-300"
-            onClick={returnToLanding}
-          >
-            <FaMicroscope size={30} className="text-violet-100" />
-            <h1 className="ml-4 text-xl font-bold text-violet-100 md:text-2xl lg:text-3xl">
-              Blog de Nandy
-            </h1>
-          </div>
-          <div className="flex pr-4 lg:hidden">
-            <Switch
-              defaultSelected
-              size="lg"
-              color="secondary"
-              startContent={<FaSun />}
-              endContent={<FaMoon />}
-              onClick={handleDarkMode}
-            />
-            <span id="nav-toggle" onClick={() => setOnOpen(!onOpen)}>
-              <GiHamburgerMenu size={30} />
+        {menuOptions.map((option, index) => (
+          <>
+            <span
+              key={option.id}
+              className={`${
+                active === index && "text-yellow-400"
+              } cursor-pointer text-sm font-medium text-gray-300 transition-colors hover:text-yellow-400`}
+              aria-current="page"
+              onClick={() => {
+                setActive(index);
+                scrollToSection(option.section);
+              }}
+            >
+              {option.name}
             </span>
-          </div>
-          <div
-            className="z-20 mt-2 hidden w-full flex-grow bg-white p-4 text-black lg:mt-0 lg:flex lg:w-auto lg:items-center lg:bg-transparent lg:p-0"
-            id="nav-content"
-          >
-            <ul className="list-reset flex-1 items-center justify-end lg:flex">
-              <li className="mr-3">
-                <span
-                  className="inline-block cursor-pointer px-4 py-2 font-bold text-white no-underline"
-                  onClick={() => handleClickOption(CASES)}
-                >
-                  Casos
-                </span>
-              </li>
-              <li className="mr-3">
-                <span
-                  className="inline-block cursor-pointer px-4 py-2 font-bold text-white no-underline"
-                  onClick={() => handleClickOption(ABOUT)}
-                >
-                  Acerca de mi
-                </span>
-              </li>
-            </ul>
-            <Switch
-              size="lg"
-              color="default"
-              startContent={<FaMoon />}
-              endContent={<FaSun />}
-              onClick={handleDarkMode}
-            />
-          </div>
-        </div>
+          </>
+        ))}
       </nav>
-      <ul
-        className={`${
-          !onOpen && "hidden"
-        } mt-12 border-y border-violet-500 bg-violet-800 p-5 text-center dark:bg-cyan-800 lg:hidden`}
+      <button
+        className="ml-auto rounded-md p-2 text-white focus:outline-none focus:ring-2 focus:ring-yellow-400 md:hidden"
+        onClick={() => setIsMenuOpen(!isMenuOpen)}
       >
-        <li
-          className="px-10 pb-3 text-2xl font-semibold text-white"
-          onClick={() => handleClickOption(CASES)}
-        >
-          Casos
-        </li>
-        <Divider />
-        <li
-          className="px-10 pt-3 text-2xl font-semibold text-white"
-          onClick={() => handleClickOption(ABOUT)}
-        >
-          Acerca de mi
-        </li>
-      </ul>
-    </>
+        {isMenuOpen ? <X className="h-6 w-6" /> : <Menu className="h-6 w-6" />}
+      </button>
+    </header>
   );
 }
