+++ src/pages/blog_client/sections/Hero.tsx
@@ -1,32 +1,54 @@
-import React from "react";
-import illus_lab from "~/assets/illus_lab.svg";
-import cell_icon from "~/assets/cell-icon.svg";
+import React, { useContext } from "react";
+import ClientContext from "~/pages/blog_client/context/ClientContext";
+import { ABOUT, CASES } from "~/constants";
+
 function Hero() {
+  const { scrollToSection } = useContext(ClientContext);
+
   return (
-    <section className="bg-gradient-to-b from-violet-700 from-70% to-violet-200 pb-16 pt-44 dark:from-cyan-700 dark:to-violet-900 lg:pt-5">
-      <div className="container mx-auto flex flex-col flex-wrap items-center px-3 md:flex-row lg:pt-20">
-        {/* <!--Left Col--> */}
-        <div className="flex w-full flex-col items-start justify-center text-center text-white md:w-1/2 md:text-left">
-          <p className="tracking-loose w-full uppercase text-white">Blog personal de Patologia</p>
-          <h1 className="my-4 text-5xl font-bold leading-tight">
-            Soy Medico Especialista en Anatomia Patologica
-          </h1>
-          <p className="mb-8 text-2xl leading-normal">
-            Los Articulos acerca de mi especialidad y otros temas de interes seran publicados Aqui
-          </p>
-          <button className="focus:shadow-outline mx-auto my-6 transform rounded-full bg-white px-8 py-4 font-bold text-gray-800 shadow-lg transition duration-300 ease-in-out hover:scale-105 hover:underline focus:outline-none lg:mx-0">
-            Contactame
-          </button>
-        </div>
-        {/* <!--Right Col--> */}
-        <div className="hidden w-full py-6 pt-44 text-center md:w-1/2 lg:flex lg:pt-6">
-          <img
-            className="w-full border-violet-500"
-            src={illus_lab}
-            alt="Ilustracion de Laboratorio"
-          />
+    <section
+      id="hero"
+      className="relative w-full overflow-hidden py-12 pt-20 md:py-24 lg:py-32 xl:py-48 xl:pt-60"
+    >
+      <div className="container relative z-10 mx-auto px-4 md:px-6">
+        <div className="grid items-center gap-6 lg:grid-cols-[2fr_1fr]">
+          <div className="flex flex-col justify-center space-y-4">
+            <div className="space-y-2">
+              <h1 className="text-3xl font-bold tracking-tighter sm:text-5xl xl:text-6xl/none">
+                Blog Profesional de Patologia
+              </h1>
+              <p className="max-w-[600px] text-gray-400 md:text-xl">
+                Publicación de casos y articulos de la Dra. Nandy Calle Peñaranda
+              </p>
+            </div>
+            <div className="flex flex-col gap-2 min-[400px]:flex-row">
+              <span
+                onClick={() => scrollToSection("cases")}
+                className="inline-flex h-10 items-center justify-center rounded-md bg-yellow-400 px-8 text-sm font-medium text-purple-950 shadow transition-colors hover:bg-yellow-500 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-yellow-500"
+              >
+                Leer el Último Caso
+              </span>
+              <span
+                onClick={() => scrollToSection("about")}
+                className="inline-flex h-10 items-center justify-center rounded-md border border-gray-700 bg-purple-950 px-8 text-sm font-medium text-white shadow-sm transition-colors hover:bg-purple-900 hover:text-gray-100 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-gray-700"
+              >
+                Acerca de Mí
+              </span>
+            </div>
+          </div>
+          <div className="relative h-[300px] lg:h-[400px]">
+            <div className="absolute inset-0 animate-pulse rounded-full bg-gradient-to-br from-yellow-400 to-purple-600"></div>
+            <img
+              alt="Jane Doe"
+              className="absolute inset-0 h-full w-full rounded-full object-cover object-center mix-blend-overlay"
+              height="400"
+              src="/placeholder.svg?height=400&width=400"
+              width="400"
+            />
+          </div>
         </div>
       </div>
+      <div className="absolute inset-0 bg-[url('/placeholder.svg?height=1080&width=1920')] bg-cover bg-center opacity-10"></div>
     </section>
   );
 }
